Extract node vulnerability conversion logic
